[hotfix][tabe, tests] Extract FunctionLookupMock to be reusable.
